Move sys-call integrity tests to benchmarks
