+++ examples/binaries/proxy/Cargo.toml
@@ -17,5 +17,6 @@ gear-wasm-builder = { path = "../../../utils/wasm-builder" }
 [lib]
 
 [features]
-std = ["codec/std", "scale-info/std"]
+wasm-wrapper = []
+std = ["wasm-wrapper", "codec/std", "scale-info/std"]
 default = ["std"]
