+++ pallets/gear/src/benchmarking/mod.rs
@@ -24,7 +24,9 @@
 mod code;
 mod sandbox;
 mod syscalls;
+mod tests;
 use syscalls::Benches;
+use tests::syscalls_integrity;
 
 use self::{
     code::{
@@ -39,7 +41,11 @@ use crate::{
     MailboxOf, Pallet as Gear, Pallet, QueueOf, Schedule,
 };
 use codec::Encode;
-use common::{benchmarking, storage::*, CodeMetadata, CodeStorage, GasPrice, GasTree, Origin};
+use common::{
+    self, benchmarking,
+    storage::{Counter, *},
+    CodeMetadata, CodeStorage, GasPrice, GasTree, Origin, QueueRunner,
+};
 use core_processor::{
     common::{DispatchOutcome, JournalNote},
     configs::BlockConfig,
@@ -80,7 +86,7 @@ const API_BENCHMARK_BATCHES: u32 = 20;
 const INSTR_BENCHMARK_BATCHES: u32 = 50;
 
 // Initializes new block.
-fn init_block<T: Config>()
+fn init_block<T: Config>(previous: Option<T::BlockNumber>)
 where
     T::AccountId: Origin,
 {
@@ -97,7 +103,7 @@ where
         )],
     };
 
-    let bn = One::one();
+    let bn = previous.unwrap_or(0u32.unique_saturated_into()) + One::one();
 
     SystemPallet::<T>::initialize(&bn, &SystemPallet::<T>::parent_hash(), &pre_digest);
     SystemPallet::<T>::set_block_number(bn);
@@ -110,7 +116,7 @@ fn process_queue<T: Config>()
 where
     T::AccountId: Origin,
 {
-    init_block::<T>();
+    init_block::<T>(None);
 
     Gear::<T>::process_queue(Default::default());
 }
@@ -260,6 +266,10 @@ benchmarks! {
         T::AccountId: Origin,
     }
 
+    check_syscalls_integrity {
+        syscalls_integrity::main_test::<T>();
+    }: {}
+
     // This bench uses `StorageMap` as a storage, due to the fact that
     // the most of the gear storages represented with this type.
     db_write_per_kb {
@@ -326,7 +336,7 @@ benchmarks! {
             None,
         ), u32::MAX.unique_saturated_into()).expect("Error during mailbox insertion");
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(RawOrigin::Signed(caller.clone()), original_message_id)
     verify {
         assert!(matches!(QueueOf::<T>::dequeue(), Ok(None)));
@@ -345,7 +355,7 @@ benchmarks! {
         let WasmModule { code, hash: code_id, .. } = WasmModule::<T>::sized(c, Location::Handle);
         let origin = RawOrigin::Signed(caller);
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(origin, code)
     verify {
         assert!(<T as pallet::Config>::CodeStorage::exists(code_id));
@@ -370,7 +380,7 @@ benchmarks! {
         <T as pallet::Config>::Currency::make_free_balance_be(&caller, caller_funding::<T>());
         let origin = RawOrigin::Signed(caller);
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(origin, code_id, salt, vec![], 100_000_000_u64, value)
     verify {
         assert!(<T as pallet::Config>::CodeStorage::exists(code_id));
@@ -398,7 +408,7 @@ benchmarks! {
         let WasmModule { code, hash, .. } = WasmModule::<T>::sized(c, Location::Handle);
         let origin = RawOrigin::Signed(caller);
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(origin, code, salt, vec![], 100_000_000_u64, value)
     verify {
         assert!(matches!(QueueOf::<T>::dequeue(), Ok(Some(_))));
@@ -414,7 +424,7 @@ benchmarks! {
         benchmarking::set_program(program_id.into_origin(), code, 1.into());
         let payload = vec![0_u8; p as usize];
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(RawOrigin::Signed(caller), program_id, payload, 100_000_000_u64, minimum_balance)
     verify {
         assert!(matches!(QueueOf::<T>::dequeue(), Ok(Some(_))));
@@ -444,7 +454,7 @@ benchmarks! {
         ), u32::MAX.unique_saturated_into()).expect("Error during mailbox insertion");
         let payload = vec![0_u8; p as usize];
 
-        init_block::<T>();
+        init_block::<T>(None);
     }: _(RawOrigin::Signed(caller.clone()), original_message_id, payload, 100_000_000_u64, minimum_balance)
     verify {
         assert!(matches!(QueueOf::<T>::dequeue(), Ok(Some(_))));
